Potential truncation in snprintf()
If a pathname is truncated, typically the application cannot present a
certificate and authentication fails as a result. However, truncation
may cause a different file to be referenced making this a potential
exploit with a carefully crafted pathname injected via getenv("HOME") or
other means.

More carefully check snprintf() return value and force failure either if
snprintf() fails or if truncation occurs. #5