enhance template file test-mail
A subject seems to only be overridable by the lib (command line switch `-s`) if there is a default subject specified in the template file. Otherwise regardless `-s` option is specified or not, the subject stays empty in the resulting mail.
Seeking for a possibility on how to hard set a subject via the lib, without having it specified in the template.